Add --this flag to set-ok and set-fail commands
With the new --this flag for set-ok and set-fail, the job is identified
by the current working directory instead of the job id. In that case,
for all jobs with Checked?=False for which the Directory column matches
the current working directory the status is set accordingly. For the
implementation of this feature, a new function for setting the status of
several jobs at once was added to the actions module.