Fixed bug where code would crash if no unchecked jobs were in the database
The recently added code to print the directory of the last unchecked job
would crash if there were no unchecked jobs in the database. This was
due to an attempt to access an empty dataframe via indices. This was
fixed by checking if unchecked jobs were found in the database.
Furthermore, if unchecked jobs are found, the database of unchecked jobs
is sorted by date to determine the latest unchecked job.